Book_Api_Flask/requirements.txt
Flask
Flask-Cors
flask-jwt-extended
Werkzeug
python-dotenv
mysql-connector-python
flask-bcrypt
